feat: transparent click when not draggable state
