+++ nattouclock/model.py
@@ -1,0 +1,4 @@
+from collections import namedtuple
+
+Point = namedtuple('Point', ['x', 'y'])
+ClockStyle = namedtuple('ClockStyle', ['size', 'color'])
